backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.9.2
